refactor: Refactor context compaction into modular pipeline
