+++ configs/configs.go
@@ -110,3 +110,8 @@ var TUITools []byte
 
 //go:embed jsons/never_open.json
 var NeverOpen []byte
+
+const (
+	PoisonRefusal     = "無法執行此操作"
+	GuardrailSentinel = "[KARAPPO]"
+)
